Backend/requirements.txt
Django==5.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
drf-yasg==1.21.7
django-csp==3.7
mysqlclient==2.2.0
Pillow==10.1.0
python-decouple==3.8
